step 1 for a clean code
