+++ ail_typo_squatting/typo.py
@@ -1,25 +1,38 @@
-import os
-import re
-import sys
-import json
-import math
-import yaml
-import string
-import inflect
-import argparse
-import random
+# Import all the modules
 
-import requests
+## The public libraries
+import os, re, sys, json, math, yaml, string, inflect, argparse, random, requests, tldextract, pathlib 
 
-import tldextract
+## The local libraries
+from conts.main import * # the constants
 
-import pathlib
-pathProg = pathlib.Path(__file__).parent.absolute()
-pathWork = ""
-for i in re.split(r"/|\\", str(pathProg))[:-1]:
-    pathWork += i + "/"
-pathEtc = pathWork + "etc"
-sys.path.append(pathEtc)
+## The typo generator
+from generator.addDash import addDash
+from generator.addDynamicDns import addDynamicDns
+from generator.addition import addition
+from generator.addTld import addTld
+from generator.changeDotDash import changeDotDash
+from generator.changeOrder import changeOrder
+from generator.commonMisspelling import commonMisspelling
+from generator.doubleReplacement import doubleReplacement
+from generator.homoglyph import homoglyph
+from generator.homophones import homophones
+from generator.missingDot import missingDot
+from generator.numeralSwap import numeralSwap
+from generator.omission import omission
+from generator.repetition import repetition
+from generator.replacement import replacement
+from generator.singularPluralize import singularPluralize
+from generator.stripDash import stripDash
+from generator.subdomain import subdomain
+from generator.vowelSwap import vowelSwap
+from generator.wrongTld import wrongTld
+from generator.wrongSld import wrongSld
+
+
+## The utils
+from generator.utils.get_pathetc import get_path_etc
+sys.path.append(get_path_etc())
 
 import pkgutil
 from typing import cast
@@ -27,57 +40,13 @@ from typing import cast
 from retrie.trie import Trie
 
 
-glyphs = {
-    '0': ['o'],
-    '1': ['l', 'i', 'ı'],
-    '2': ['ƻ'],
-    '5': ['ƽ'],
-    'a': ['à', 'á', 'à', 'â', 'ã', 'ä', 'å', 'ɑ', 'ạ', 'ǎ', 'ă', 'ȧ', 'ą', 'ə'],
-    'b': ['d', 'ʙ', 'ɓ', 'ḃ', 'ḅ', 'ḇ', 'ƅ'],
-    'c': ['e', 'ƈ', 'ċ', 'ć', 'ç', 'č', 'ĉ', 'ᴄ'],
-    'd': ['b', 'cl', 'ɗ', 'đ', 'ď', 'ɖ', 'ḑ', 'ḋ', 'ḍ', 'ḏ', 'ḓ'],
-    'e': ['c', 'é', 'è', 'ê', 'ë', 'ē', 'ĕ', 'ě', 'ė', 'ẹ', 'ę', 'ȩ', 'ɇ', 'ḛ'],
-    'f': ['ƒ', 'ḟ'],
-    'g': ['q', 'ɢ', 'ɡ', 'ġ', 'ğ', 'ǵ', 'ģ', 'ĝ', 'ǧ', 'ǥ'],
-    'h': ['ĥ', 'ȟ', 'ħ', 'ɦ', 'ḧ', 'ḩ', 'ⱨ', 'ḣ', 'ḥ', 'ḫ', 'ẖ'],
-    'i': ['1', 'l', 'í', 'ì', 'ï', 'ı', 'ɩ', 'ǐ', 'ĭ', 'ỉ', 'ị', 'ɨ', 'ȋ', 'ī', 'ɪ'],
-    'j': ['ʝ', 'ǰ', 'ɉ', 'ĵ'],
-    'k': ['lc', 'ḳ', 'ḵ', 'ⱪ', 'ķ', 'ᴋ'],
-    'l': ['1', 'i', 'ɫ', 'ł', 'ı', 'ɩ'],
-    'm': ['n', 'nn', 'rn', 'rr', 'ṁ', 'ṃ', 'ᴍ', 'ɱ', 'ḿ'],
-    'n': ['m', 'r', 'ń', 'ṅ', 'ṇ', 'ṉ', 'ñ', 'ņ', 'ǹ', 'ň', 'ꞑ'],
-    'o': ['0', 'ȯ', 'ọ', 'ỏ', 'ơ', 'ó', 'ö', 'ᴏ'],
-    'p': ['ƿ', 'ƥ', 'ṕ', 'ṗ'],
-    'q': ['g', 'ʠ'],
-    'r': ['ʀ', 'ɼ', 'ɽ', 'ŕ', 'ŗ', 'ř', 'ɍ', 'ɾ', 'ȓ', 'ȑ', 'ṙ', 'ṛ', 'ṟ'],
-    's': ['ʂ', 'ś', 'ṣ', 'ṡ', 'ș', 'ŝ', 'š', 'ꜱ'],
-    't': ['ţ', 'ŧ', 'ṫ', 'ṭ', 'ț', 'ƫ'],
-    'u': ['ᴜ', 'ǔ', 'ŭ', 'ü', 'ʉ', 'ù', 'ú', 'û', 'ũ', 'ū', 'ų', 'ư', 'ů', 'ű', 'ȕ', 'ȗ', 'ụ'],
-    'v': ['ṿ', 'ⱱ', 'ᶌ', 'ṽ', 'ⱴ', 'ᴠ'],
-    'w': ['vv', 'ŵ', 'ẁ', 'ẃ', 'ẅ', 'ⱳ', 'ẇ', 'ẉ', 'ẘ', 'ᴡ'],
-    'x': ['ẋ', 'ẍ'],
-    'y': ['ʏ', 'ý', 'ÿ', 'ŷ', 'ƴ', 'ȳ', 'ɏ', 'ỿ', 'ẏ', 'ỵ'],
-    'z': ['ʐ', 'ż', 'ź', 'ᴢ', 'ƶ', 'ẓ', 'ẕ', 'ⱬ']
-    }
 
-numerals = [
-    ["0", "zero"],
-    ["1", "one", "first"],
-    ["2", "two", "second"],
-    ["3", "three", "third"],
-    ["4", "four", "fourth", "for"],
-    ["5", "five", "fifth"],
-    ["6", "six", "sixth"],
-    ["7", "seven", "seventh"],
-    ["8", "eight", "eighth"],
-    ["9", "nine", "ninth"]
-]
-
-algo_list = ["omission", "repetition", "changeOrder", "replacement", "doubleReplacement", "addition", "missingDot", "stripDash", "vowelSwap", "addDash", "homoglyph", "commonMisspelling", "homophones", "wrongTld", "addTld", "subdomain", "singularPluralize", "changeDotDash", "wrongSld", "numeralSwap", "addDynamicDns"]
-
-type_request = ['A', 'AAAA', 'NS', 'MX']
-
-exclude_tld = ["gov.pl"]
+# Import all the constants of data from the file conts/main.py
+# If you wanna add a new algorithm, you have to add it in the list algo_list
+numerals = const_get_numeral()
+algo_list = const_get_algo_name_list()
+type_request = const_get_dns_request_type()
+exclude_tld = const_get_excluded_tld()
 
 
 def globalAppend(loclist):
@@ -112,7 +81,6 @@ def globalAppend(loclist):
 
     return r[-1]
     
-
 def checkResult(resultLoc, resultList, givevariations, algoName=''):
     """
     Verify if element in resultLoc not exist in resultList before adding them in resultList
@@ -200,1028 +168,7 @@ def final_treatment(domain, resultList, limit, givevariations, keeporiginal, alg
 
     return resultList
 
-
-def omission(domain, resultList, verbose, limit, givevariations=False,  keeporiginal=False, combo=False):
-    """Leave out a letter of the domain name"""
-
-    if not len(resultList) >= limit:
-        if verbose:
-            print("[+] Omission")
-
-        resultLoc = list()
-        loclist = list()
-
-        prefix, domain_without_tld, tld = parse_domain(domain)
-        domainList = [domain_without_tld]
-
-        for name in domainList:
-            for i in range(0,len(name)):
-                resultLoc.append(prefix + name)
-                loc = prefix + name[0:i]
-                loc += name[i+1:len(name)]
-
-                if loc not in resultLoc:
-                    resultLoc.append(loc)
-
-            if resultLoc:
-                loclist.append(resultLoc)
-                resultLoc = list()
-
-        loclist.append([tld])
-
-        rLoc = globalAppend(loclist)
-
-        if verbose:
-            print(f"{len(rLoc)}\n")
-
-        if combo:
-            rLoc = checkResult(rLoc, resultList, givevariations, "omission")
-            rLoc = final_treatment(domain, rLoc, limit, givevariations, keeporiginal, "omission")
-            return rLoc
-
-        resultList = checkResult(rLoc, resultList, givevariations, "omission")
-        resultList = final_treatment(domain, resultList, limit, givevariations, keeporiginal, "omission")
-
-    return resultList
-
-def repetition(domain, resultList, verbose, limit, givevariations=False,  keeporiginal=False, combo=False):
-    """Characters Repetition"""
-
-    if not len(resultList) >= limit:
-        if verbose:
-            print("[+] Characters Repetition")
-
-        resultLoc = list()
-        loclist = list()
-
-        prefix, domain_without_tld, tld = parse_domain(domain)
-        domainList = [domain_without_tld]
-
-        for name in domainList:
-            resultLoc.append(prefix + name)
-            for i, c in enumerate(name):
-                if prefix + name[:i] + c + name[i:] not in resultLoc:
-                    resultLoc.append(prefix + name[:i] + c + name[i:])
-
-            if resultLoc:
-                loclist.append(resultLoc)
-                resultLoc = list()
-
-        loclist.append([tld])
-        rLoc = globalAppend(loclist)
-
-        if verbose:
-            print(f"{len(rLoc)}\n")
-
-        if combo:
-            rLoc = checkResult(rLoc, resultList, givevariations, 'repetition')
-            rLoc = final_treatment(domain, rLoc, limit, givevariations, keeporiginal, "repetition")
-            return rLoc
-
-        resultList = checkResult(rLoc, resultList, givevariations, 'repetition')
-        resultList = final_treatment(domain, resultList, limit, givevariations, keeporiginal, "repetition")
-
-    return resultList
-
-def changeOrder(domain, resultList, verbose, limit, givevariations=False,  keeporiginal=False, combo=False):
-    """Change the order of letters in word"""
-
-    if not len(resultList) >= limit:
-        if verbose:
-            print("[+] Change Order")
-
-        resultLoc = list()
-        loclist = list()
-
-        prefix, domain_without_tld, tld = parse_domain(domain)
-        domainList = [domain_without_tld]
-
-        for name in domainList:
-            if len(name) == 1:
-                resultLoc.append(prefix + name)
-            else:
-                for i in range(0, len(name)):
-                    loc = name[0:i] + name[i+1:]
-                    for j in range(0, len(loc)):
-                        inter = prefix + loc[:j] + name[i] + loc[j:]
-                        if not inter in resultLoc:
-                            resultLoc.append(inter)
-
-            if resultLoc:
-                loclist.append(resultLoc)
-                resultLoc = list()
-
-        loclist.append([tld])
-        rLoc = globalAppend(loclist)
-
-        if verbose:
-            print(f"{len(rLoc)}\n")
-        
-        if combo:
-            rLoc = checkResult(rLoc, resultList, givevariations, 'changeOrder')
-            # print("Hello")
-            rLoc = final_treatment(domain, rLoc, limit, givevariations, keeporiginal, "changeOrder")
-            return rLoc
-
-        resultList = checkResult(rLoc, resultList, givevariations, 'changeOrder')
-        resultList = final_treatment(domain, resultList, limit, givevariations, keeporiginal, "changeOrder")
-
-    return resultList
-
-
-def replacement(domain, resultList, verbose, limit, givevariations=False,  keeporiginal=False, combo=False):
-    """Adjacent character replacement to the immediate left and right on the keyboard"""
-
-    if not len(resultList) >= limit:
-        if verbose:
-            print("[+] Replacement")
-
-        resultLoc = list()
-        loclist = list()
-
-        prefix, domain_without_tld, tld = parse_domain(domain)
-        domainList = [domain_without_tld]
-
-        for name in domainList:
-            for i in (*range(48, 58), *range(97, 123)):
-                for j in range(0, len(name)):
-                    pre = name[:j]
-                    suf = name[j+1:]
-                    variation = prefix + pre + chr(i) + suf
-                    if variation not in resultLoc:
-                        resultLoc.append(variation)
-
-            if resultLoc:
-                loclist.append(resultLoc)
-                resultLoc = list()
-
-        loclist.append([tld])
-        rLoc = globalAppend(loclist)
-
-        if verbose:
-            print(f"{len(rLoc)}\n")
-
-        if combo:
-            rLoc = checkResult(rLoc, resultList, givevariations, 'replacement')
-            rLoc = final_treatment(domain, rLoc, limit, givevariations, keeporiginal, "replacement")
-            return rLoc
-
-        resultList = checkResult(rLoc, resultList, givevariations, 'replacement')
-        resultList = final_treatment(domain, resultList, limit, givevariations, keeporiginal, "replacement")
-
-    return resultList
-
-
-def doubleReplacement(domain, resultList, verbose, limit, givevariations=False,  keeporiginal=False, combo=False):
-    """Double Character Replacement"""
-
-    if not len(resultList) >= limit:
-        if verbose:
-            print("[+] Double Character Replacement")
-
-        resultLoc = list()
-        loclist = list()
-
-        prefix, domain_without_tld, tld = parse_domain(domain)
-        domainList = [domain_without_tld]
-
-        for name in domainList:
-            for i in (*range(48, 58), *range(97, 123)):
-                for j in range(0, len(name)):
-                    pre = name[:j]
-                    suf = name[j+2:]
-                    variation = prefix + pre + chr(i) + chr(i) + suf
-                    if variation not in resultLoc:
-                        resultLoc.append(variation)
-
-            if resultLoc:
-                loclist.append(resultLoc)
-                resultLoc = list()
-
-        loclist.append([tld])
-        rLoc = globalAppend(loclist)
-
-        if verbose:
-            print(f"{len(rLoc)}\n")
-
-        if combo:
-            rLoc = checkResult(rLoc, resultList, givevariations, 'doubleReplacement')
-            rLoc = final_treatment(domain, rLoc, limit, givevariations, keeporiginal, "doubleReplacement")
-            return rLoc
-
-        resultList = checkResult(rLoc, resultList, givevariations, 'doubleReplacement')
-        resultList = final_treatment(domain, resultList, limit, givevariations, keeporiginal, "doubleReplacement")
-
-    return resultList
-
-
-
-def addition(domain, resultList, verbose, limit, givevariations=False,  keeporiginal=False, combo=False):
-    """Add a character in the domain name"""
-
-    if not len(resultList) >= limit:
-        if verbose:
-            print("[+] Addition")
-
-        resultLoc = list()
-        loclist = list()
-
-        prefix, domain_without_tld, tld = parse_domain(domain)
-        domainList = [domain_without_tld]
-
-        for name in domainList:
-            for i in (*range(48, 58), *range(97, 123)):
-                # Adding 'i' in front of 'name'
-                variation = prefix + chr(i) + name
-                if variation not in resultLoc:
-                    resultLoc.append(variation)
-
-                # Adding 'i' at the end of 'name'
-                variation = prefix + name + chr(i)
-                if variation not in resultLoc:
-                    resultLoc.append(variation)
-
-                for j in range(0, len(name)):
-                    variation = prefix + name[:j] + chr(i) + name[j:]
-                    if variation not in resultLoc:
-                        resultLoc.append(variation)
-
-            if resultLoc:
-                loclist.append(resultLoc)
-                resultLoc = list()
-
-        loclist.append([tld])
-
-        rLoc = globalAppend(loclist)
-
-        if verbose:
-            print(f"{len(rLoc)}\n")
-
-        if combo:
-            rLoc = checkResult(rLoc, resultList, givevariations, 'addition')
-            rLoc = final_treatment(domain, rLoc, limit, givevariations, keeporiginal, "addition")
-            return rLoc
-
-        resultList = checkResult(rLoc, resultList, givevariations, 'addition')
-        resultList = final_treatment(domain, resultList, limit, givevariations, keeporiginal, "addition")
-
-    return resultList
-
-
-def utilMissingDot(resultLoc, loc):
-    """Function for missingDot algorithm"""
-    
-    i = 0
-    while "." in loc:
-        loc2 = loc[::-1].replace(".", "", 1)[::-1]
-
-        loc = loc.replace(".", "", 1)
-
-        if loc not in resultLoc:
-            resultLoc.append(loc)
-
-        if loc2 not in resultLoc:
-            resultLoc.append(loc2) 
-        i += 1
-        
-    return resultLoc
-
-def missingDot(domain, resultList, verbose, limit, givevariations=False,  keeporiginal=False, combo=False):
-    """Delete a dot from the domain name"""
-
-    if not len(resultList) >= limit:
-        if verbose:
-            print("[+] Missing Dot")
-
-        resultLoc = list()
-        cp = 0
-
-        domainList = domain.split(".")
-
-        loc = domain
-        utilMissingDot(resultLoc, loc)
-        
-        loc = f"www{domain}"
-        utilMissingDot(resultLoc, loc)
-
-        loc_result_list = resultList.copy()
-
-        for i in range(0, len(resultLoc)):
-            if domainList[-1] in resultLoc[i].split(".")[0]:
-                resultLoc[i] = resultLoc[i] + ".com"
-
-            if givevariations:
-                flag = False
-                for var in algo_list:
-                    if [resultLoc[i], var] in resultList:
-                        flag = True
-                if not flag:
-                    cp += 1
-                    loc_result_list.append([resultLoc[i], "missingDot"])
-            elif resultLoc[i] not in resultList:
-                cp += 1
-                loc_result_list.append(resultLoc[i])             
-
-        if verbose:
-            print(f"{cp}\n")
-
-        # if combo:
-        #     return final_treatment(domain, loc_result_list, limit, givevariations, keeporiginal, "missingDot")
-        return final_treatment(domain, loc_result_list, limit, givevariations, keeporiginal, "missingDot")
-
-
-    return resultList
-
-def stripDash(domain, resultList, verbose, limit, givevariations=False,  keeporiginal=False, combo=False):
-    """Delete a dash from the domain name"""
-
-    if not len(resultList) >= limit:
-        if verbose:
-            print("[+] Strip Dash")
-
-        loc = domain
-        cp = 0
-        loc_result_list = resultList.copy()
-        while "-" in loc:
-            loc2 = loc[::-1].replace("-", "", 1)[::-1]
-            loc = loc.replace("-", "", 1)
-
-            if givevariations:
-                flag = False
-                for var in algo_list:
-                    if [loc, var] in resultList:
-                        flag = True
-                if not flag:
-                    cp += 1
-                    loc_result_list.append([loc, "stripDash"])
-
-                flag = False
-                for var in algo_list:
-                    if [loc2, var] in resultList:
-                        flag = True
-                if not flag:
-                    cp += 1
-                    loc_result_list.append([loc2, "stripDash"])
-            else:
-                if loc not in resultList:
-                    cp += 1
-                    loc_result_list.append(loc)
-
-                if loc2 not in resultList:
-                    cp += 1
-                    loc_result_list.append(loc2)
-        
-        if verbose:
-            print(f"{cp}\n")
-
-        return final_treatment(domain, loc_result_list, limit, givevariations, keeporiginal, "stripDash")
-
-    return resultList
-
-def vowelSwap(domain, resultList, verbose, limit, givevariations=False,  keeporiginal=False, combo=False):
-    """Swap vowels within the domain name"""
-
-    if not len(resultList) >= limit:
-        if verbose:
-            print("[+] Vowel Swap")
-
-        resultLoc = list()
-        loclist = list()
-        # vowels = 'aeiouy'
-        vowels = ["a", "e", "i", "o", "u", "y"]
-
-        prefix, domain_without_tld, tld = parse_domain(domain)
-        domainList = [domain_without_tld]
-
-        for name in domainList:
-            for j in vowels:
-                for k in vowels:
-                    if j != k:
-                        loc = prefix + name.replace(k, j)
-                        if loc not in resultLoc:
-                            resultLoc.append(loc)
-
-            if resultLoc:
-                loclist.append(resultLoc)
-                resultLoc = list()
-
-        loclist.append([tld])
-        rLoc = globalAppend(loclist)
-
-        if verbose:
-            print(f"{len(rLoc)}\n")
-
-        if combo:
-            rLoc = checkResult(rLoc, resultList, givevariations, "vowelSwap")
-            rLoc = final_treatment(domain, rLoc, limit, givevariations, keeporiginal, "vowelSwap")
-            return rLoc
-
-        resultList = checkResult(rLoc, resultList, givevariations, "vowelSwap")
-        resultList = final_treatment(domain, resultList, limit, givevariations, keeporiginal, "vowelSwap")
-
-    return resultList
-
-
-def addDash(domain, resultList, verbose, limit, givevariations=False,  keeporiginal=False, combo=False):
-    """Add a dash between the first and last character in a string"""
-
-    if not len(resultList) >= limit:
-        if verbose:
-            print("[+] Add Dash")
-
-        resultLoc = list()
-        loclist = list()
-
-        prefix, domain_without_tld, tld = parse_domain(domain)
-        domainList = [domain_without_tld]
-
-        for name in domainList:
-            if len(name) == 1:
-                resultLoc.append(prefix + name)
-            else:
-                resultLoc.append(prefix + name)
-                for i in range(1, len(name)):
-                    if prefix + name[:i] + '-' + name[i:] not in resultLoc:
-                        resultLoc.append(prefix + name[:i] + '-' + name[i:])
-
-            if resultLoc:
-                loclist.append(resultLoc)
-                resultLoc = list()
-
-        loclist.append([tld])
-        rLoc = globalAppend(loclist)
-
-        if verbose:
-            print(f"{len(rLoc)}\n")
-
-        if combo:
-            rLoc = checkResult(rLoc, resultList, givevariations, "addDash")
-            rLoc = final_treatment(domain, rLoc, limit, givevariations, keeporiginal, "addDash")
-            return rLoc
-
-        resultList = checkResult(rLoc, resultList, givevariations, "addDash")
-        resultList = final_treatment(domain, resultList, limit, givevariations, keeporiginal, "addDash")
-
-    return resultList
-
-
-def homoglyph(domain, resultList, verbose, limit, givevariations=False,  keeporiginal=False, all=False, combo=False):
-    """One or more characters that look similar to another character but are different are called homogylphs"""
-
-    if not len(resultList) >= limit:
-        if verbose:
-            print("[+] Homoglyph")
-
-        def mix(domain):
-            for w in range(1, len(domain)):
-                for i in range(len(domain)-w+1):
-                    pre = domain[:i]
-                    win = domain[i:i+w]
-                    suf = domain[i+w:]
-                    for c in win:
-                        for g in glyphs.get(c, []):
-                            yield pre + win.replace(c, g) + suf
-
-        prefix, domain_without_tld, tld = parse_domain(domain)
-        domainList = [domain_without_tld]
-
-        s = '.'.join(domainList)
-
-        result1 = set(mix(s))
-        result2 = set()
-        cp = 0
-        loc_result_list = resultList.copy()
-
-        if all:
-            for r in result1:
-                result2.update(set(mix(r)))
-        
-        for element in list(result1 | result2):
-            element = prefix + element + '.' + tld
-            if givevariations:
-                flag = False
-                for var in algo_list:
-                    if [element, var] in resultList:
-                        flag = True
-                if not flag:
-                    cp += 1
-                    loc_result_list.append([element, "homoglyph"])
-
-            elif element not in resultList:
-                cp += 1
-                loc_result_list.append(element)
-
-        if verbose:
-            print(f"{cp}\n")
-
-        return final_treatment(domain, loc_result_list, limit, givevariations, keeporiginal, "homoglyph")
-
-    return resultList
-
-
-def commonMisspelling(domain, resultList, verbose, limit, givevariations=False,  keeporiginal=False, combo=False):
-    """Change a word by is misspellings"""
-    # https://en.wikipedia.org/wiki/Wikipedia:Lists_of_common_misspellings/For_machines
-
-    if not len(resultList) >= limit:
-        if verbose:
-            print("[+] Common Misspelling")
-
-        with open(pathEtc + "/common-misspellings.json", "r") as read_json:
-            misspelling = json.load(read_json)
-            keys = misspelling.keys()
-
-        prefix, domain_without_tld, tld = parse_domain(domain)
-        domainList = [domain_without_tld]
-
-        resultLoc = list()
-        loclist = list()
-
-        for name in domainList:
-            if name in keys:
-                misspell = misspelling[name].split(",")
-                for mis in misspell:
-                    if prefix + mis.replace(" ","") not in resultLoc:
-                        resultLoc.append(prefix + mis.replace(" ",""))
-            elif name not in resultLoc:
-                resultLoc.append(prefix + name)
-
-            if resultLoc:
-                loclist.append(resultLoc)
-                resultLoc = list()
-
-        loclist.append([tld])
-        rLoc = globalAppend(loclist)
-
-        if verbose:
-            print(f"{len(rLoc)}\n")
-
-        if combo:
-            rLoc = checkResult(rLoc, resultList, givevariations, "commonMisspelling")
-            rLoc = final_treatment(domain, rLoc, limit, givevariations, keeporiginal, "commonMisspelling")
-            return rLoc
-
-        resultList = checkResult(rLoc, resultList, givevariations, "commonMisspelling")
-        resultList = final_treatment(domain, resultList, limit, givevariations, keeporiginal, "commonMisspelling")
-
-    return resultList
-
-
-def homophones(domain, resultList, verbose, limit, givevariations=False,  keeporiginal=False, combo=False):
-    """Change word by an other who sound the same when spoken"""
-    # From http://en.wikipedia.org/wiki/Wikipedia:Lists_of_common_misspellings/Homophones
-    # Last updated 04/2020
-    # cat /tmp/h | sed 's/^[ ]*//g' | egrep -v "These pairs become homophones in certain dialects only|^Names" | sed -E 's/ (and|or) /,/g' | sed 's/\//,/g' | sed 's/,,/,/g' | tr '[:upper:]' '[:lower:]' | tr -d " '" | grep -v "^$"
-
-    if not len(resultList) >= limit:
-        if verbose:
-            print("[+] Homophones")
-
-        with open(pathEtc + "/homophones.txt", "r") as read_file:
-            homophones = read_file.readlines()
-        
-        prefix, domain_without_tld, tld = parse_domain(domain)
-        domainList = [domain_without_tld]
-
-        resultLoc = list()
-        loclist = list()
-
-        for name in domainList:
-            for lines in homophones:
-                line = lines.split(",")
-                for word in line:
-                    if name == word.rstrip("\n"):
-                        for otherword in line:
-                            if prefix + otherword.rstrip("\n") not in resultLoc and otherword.rstrip("\n") != name:
-                                resultLoc.append(prefix + otherword.rstrip("\n"))
-                    elif name not in resultLoc:
-                        resultLoc.append(prefix + name)
-
-            if resultLoc:
-                loclist.append(resultLoc)
-                resultLoc = list()
-
-        loclist.append([tld])
-        rLoc = globalAppend(loclist)
-
-        if verbose:
-            print(f"{len(rLoc)}\n")
-
-        if combo:
-            rLoc = checkResult(rLoc, resultList, givevariations, "homophones")
-            rLoc = final_treatment(domain, rLoc, limit, givevariations, keeporiginal, "homophones")
-            return rLoc
-
-        resultList = checkResult(rLoc, resultList, givevariations, "homophones")
-        resultList = final_treatment(domain, resultList, limit, givevariations, keeporiginal, "homophones")
-
-    return resultList
-    
-
-def wrongTld(domain, resultList, verbose, limit, givevariations=False, keeporiginal=False, combo=False):
-    """Change the original top level domain to another"""
-    # https://data.iana.org/TLD/tlds-alpha-by-domain.txt
-    # Version 2022102502
-
-    if not len(resultList) >= limit:
-        if verbose:
-            print("[+] Wrong Tld")
-
-        with open(pathEtc + "/tlds-alpha-by-domain.txt", "r") as read_file:
-            tlds = read_file.readlines()
-        
-        originalTld = domain.split(".")[-1]
-
-        prefix, domain_without_tld, possible_complete_tld = parse_domain(domain)
-
-        domainLoc = ""
-        domainLoc_orig = ''
-        cp = 0
-        loc_result_list = resultList.copy()
-
-        if possible_complete_tld != originalTld:
-            if prefix:
-                domainLoc = prefix
-
-            for element in domain_without_tld.split("."):
-                domainLoc += element + "."
-        for element in domain.split(".")[:-1]:
-            domainLoc_orig += element + "."
-
-        for tld in tlds:
-            if tld.lower().rstrip("\n") != originalTld:
-                cp += 1
-                if givevariations:
-                    if possible_complete_tld != originalTld:
-                        cp += 1
-                        loc_result_list.append([domainLoc + tld.lower().rstrip("\n"), 'wrongTld'])
-                    loc_result_list.append([domainLoc_orig + tld.lower().rstrip("\n"), 'wrongTld'])
-                else:
-                    if possible_complete_tld != originalTld:
-                        cp += 1
-                        loc_result_list.append(domainLoc + tld.lower().rstrip("\n"))
-                    loc_result_list.append(domainLoc_orig + tld.lower().rstrip("\n"))
-        
-        maybe_pkg_data = pkgutil.get_data("tldextract", ".tld_set_snapshot")
-        pkg_data = cast(bytes, maybe_pkg_data)
-        text = pkg_data.decode("utf-8")
-        public, private = tldextract.suffix_list.extract_tlds_from_suffix_list(text)
-
-        if possible_complete_tld == originalTld:
-            split_tld = possible_complete_tld.split(".")
-            for p in public:
-                if p.endswith("." + split_tld[-1]):
-                    loc = p.split(".")[-2] + "." + p.split(".")[-1]
-                    if givevariations:
-                        if not [domainLoc_orig + loc, "wrongTld"] in resultList:
-                            loc_result_list.append([domainLoc_orig + loc, "wrongTld"])
-                    else:
-                        if not domainLoc_orig + loc in resultList:
-                            loc_result_list.append(domainLoc_orig + loc)
-
-            for p in private:
-                if p.endswith("." + split_tld[-1]):
-                    loc = p.split(".")[-2] + "." + p.split(".")[-1]
-                    if givevariations:
-                        if not [domainLoc_orig + loc, "wrongTld"] in resultList:
-                            cp += 1
-                            loc_result_list.append([domainLoc_orig + loc, "wrongTld"])
-                    else:
-                        if not domainLoc_orig + loc in resultList:
-                            cp += 1
-                            loc_result_list.append(domainLoc_orig + loc)
-
-        if verbose:
-            print(f"{cp}\n")
-
-        return final_treatment(domain, loc_result_list, limit, givevariations, keeporiginal, "wrongTld")
-
-    return resultList
-
-def wrongSld(domain, resultList, verbose, limit, givevariations=False, keeporiginal=False, combo=False):
-    """Change the original second level domain to another"""
-
-    if not len(resultList) >= limit:
-        if verbose:
-            print("[+] Wrong Sld")
-        
-        originalTld = domain.split(".")[-1]
-
-        prefix, domain_without_tld, possible_complete_tld = parse_domain(domain)
-
-        domainLoc = ""
-        cp = 0
-        loc_result_list = resultList.copy()
-
-        if possible_complete_tld != originalTld:
-            if prefix:
-                domainLoc = prefix
-
-            for element in domain_without_tld.split("."):
-                domainLoc += element + "."
-        
-        maybe_pkg_data = pkgutil.get_data("tldextract", ".tld_set_snapshot")
-        pkg_data = cast(bytes, maybe_pkg_data)
-        text = pkg_data.decode("utf-8")
-        public, private = tldextract.suffix_list.extract_tlds_from_suffix_list(text)
-
-        if possible_complete_tld != originalTld:
-            split_tld = possible_complete_tld.split(".")
-            for p in public:
-                if p.endswith("." + split_tld[-1]):
-                    loc = p.split(".")[-2] + "." + p.split(".")[-1]
-                    if givevariations:
-                        if not [domainLoc + loc, "wrongSld"] in resultList:
-                            loc_result_list.append([domainLoc + loc, "wrongSld"])
-                    else:
-                        if not domainLoc + loc in resultList:
-                            loc_result_list.append(domainLoc + loc)
-
-            for p in private:
-                if p.endswith("." + split_tld[-1]):
-                    loc = p.split(".")[-2] + "." + p.split(".")[-1]
-                    if givevariations:
-                        if not [domainLoc + loc, "wrongSld"] in resultList:
-                            cp += 1
-                            loc_result_list.append([domainLoc + loc, "wrongSld"])
-                    else:
-                        if not domainLoc + loc in resultList:
-                            cp += 1
-                            loc_result_list.append(domainLoc + loc)
-
-        if verbose:
-            print(f"{cp}\n")
-    
-        return final_treatment(domain, loc_result_list, limit, givevariations, keeporiginal, "wrongSld")
-
-    return resultList
-
-
-def addTld(domain, resultList, verbose, limit, givevariations=False, keeporiginal=False, combo=False):
-    """Adding a tld before the original tld"""
-    # https://data.iana.org/TLD/tlds-alpha-by-domain.txt
-    # Version 2022012800
-
-    if not len(resultList) >= limit:
-        if verbose:
-            print("[+] Adding Tld")
-        with open(pathEtc + "/tlds-alpha-by-domain.txt", "r") as read_file:
-            tlds = read_file.readlines()
-        
-        cp = 0 
-        loc_result_list = resultList.copy()
-        for tld in tlds:
-            cp += 1
-            if givevariations:
-                loc_result_list.append([domain + "." + tld.lower().rstrip("\n"), 'addTld'])
-            else:
-                loc_result_list.append(domain + "." + tld.lower().rstrip("\n"))
-
-        if verbose:
-            print(f"{cp}\n")
-
-        # if combo:
-        #     return final_treatment(domain, loc_result_list, limit, givevariations, keeporiginal, "addTld")
-
-        # resultList = resultList + loc_result_list
-        # resultList = final_treatment(domain, resultList, limit, givevariations, keeporiginal, "addTld")
-
-        return final_treatment(domain, loc_result_list, limit, givevariations, keeporiginal, "addTld")
-
-    return resultList
-
-
-def subdomain(domain, resultList, verbose, limit, givevariations=False,  keeporiginal=False, combo=False):
-    """Insert a dot at varying positions to create subdomain"""
-
-    if not len(resultList) >= limit:
-        if verbose:
-            print("[+] Subdomain")
-
-        cp = 0
-        loc_result_list = resultList.copy()
-
-        prefix, domain_without_tld, domain_tld = parse_domain(domain)
-        
-        if prefix:
-            domain_tmp = prefix + domain_without_tld
-        else:
-            domain_tmp = domain_without_tld
-
-        for i in range(1, len(domain_tmp)):
-            if domain_tmp[i] not in ['-', '.'] and domain_tmp[i-1] not in ['-', '.']:
-                cp += 1
-                if givevariations:
-                    loc_result_list.append([domain_tmp[:i] + '.' + domain_tmp[i:] + '.' + domain_tld, 'subdomain'])
-                else:
-                    loc_result_list.append(domain_tmp[:i] + '.' + domain_tmp[i:] + '.' + domain_tld)
-
-        if verbose:
-            print(f"{cp}\n")
-
-        return final_treatment(domain, loc_result_list, limit, givevariations, keeporiginal, "subdomain")
-
-    return resultList
-
-
-def singularPluralize(domain, resultList, verbose, limit, givevariations=False, keeporiginal=False, combo=False):
-    """Create by making a singular domain plural and vice versa"""
-
-    if not len(resultList) >= limit:
-        if verbose:
-            print("[+] Singular Pluralize")
-            
-        resultLoc = list()
-        loclist = list()
-        inflector = inflect.engine()
-
-        prefix, domain_without_tld, tld = parse_domain(domain)
-        domainList = [domain_without_tld]
-
-        for name in domainList:
-            loc = inflector.plural(name)
-            resultLoc.append(prefix + name)
-
-            if loc and loc not in resultLoc:
-                resultLoc.append(prefix+ loc)
-            
-            if resultLoc:
-                loclist.append(resultLoc)
-                resultLoc = list()
-
-        loclist.append([tld])
-        rLoc = globalAppend(loclist)
-
-        if verbose:
-            print(f"{len(rLoc)}\n")
-
-        if combo:
-            rLoc = checkResult(rLoc, resultList, givevariations, "singularPluralize")
-            rLoc = final_treatment(domain, rLoc, limit, givevariations, keeporiginal, "singularPluralize")
-            return rLoc
-
-        resultList = checkResult(rLoc, resultList, givevariations, "singularPluralize")
-        resultList = final_treatment(domain, resultList, limit, givevariations, keeporiginal, "singularPluralize")
-
-    return resultList
-
-def changeDotDash(domain, resultList, verbose, limit, givevariations=False,  keeporiginal=False, combo=False):
-    """Change dot to dash"""
-
-    if not len(resultList) >= limit:
-        if verbose:
-            print("[+] Change dot to dash")
-
-        domain_extract = tldextract.extract(domain)
-        original_tld = domain_extract.suffix
-        original_dwt = domain_extract.domain
-
-        if not original_tld:
-            print("[-] Domain not valid")
-            exit(-1)
-
-        domain_list = [original_dwt]
-
-        loc = domain
-        loc2 = domain
-        loc_result_list = resultList.copy()
-        while "." in loc:
-            resultLoc = list()
-            resultLoc2 = list()
-            loc2 = loc2[::-1].replace(".", "-", 1)[::-1]
-            loc = loc.replace(".", "-", 1)
-
-            loc_domain_extract = tldextract.extract(loc)
-            loc_dwt = loc_domain_extract.domain
-            loc_tld = loc_domain_extract.suffix
-
-            loc2_domain_extract = tldextract.extract(loc2)
-            loc2_dwt = loc2_domain_extract.domain
-            loc2_tld = loc2_domain_extract.suffix
-
-
-            if "." not in loc and loc_dwt not in domain_list:
-                resultLoc = addTld(loc, resultLoc, verbose, limit, givevariations)
-                loc_result_list = checkResult(resultLoc, resultList, givevariations, "changeDotDash")
-                domain_list.append(loc_dwt)
-            elif loc_dwt not in domain_list:
-                if loc_tld != original_tld:
-                    resultLoc = addTld(loc, resultLoc, verbose, limit, givevariations)
-                    loc_result_list = checkResult(resultLoc, resultList, givevariations, "changeDotDash")
-                else:
-                    if givevariations:
-                        loc_result_list.append([loc, 'changeDotDash'])
-                    else:
-                        loc_result_list.append(loc)
-                domain_list.append(loc_dwt)
-
-            if "." not in loc2 and loc2_dwt not in domain_list:
-                resultLoc2 = addTld(loc2, resultLoc2, verbose, limit, givevariations)
-                loc_result_list = checkResult(resultLoc2, resultList, givevariations, "changeDotDash")
-                domain_list.append(loc2_dwt)
-            elif loc2_tld != original_tld:
-                if loc2_dwt not in domain_list:
-                    resultLoc2 = addTld(loc2, resultLoc2, verbose, limit, givevariations)
-                    loc_result_list = checkResult(resultLoc2, resultList, givevariations, "changeDotDash")
-                else:
-                    if givevariations:
-                        loc_result_list.append([loc2, 'changeDotDash'])
-                    else:
-                        loc_result_list.append(loc2)
-                domain_list.append(loc2_dwt)
-
-        return final_treatment(domain, loc_result_list, limit, givevariations, keeporiginal, "changeDotDash")
-
-    return resultList
-
-
-def addDynamicDns(domain, resultList, verbose, limit, givevariations=False,  keeporiginal=False, combo=False):
-    """Add dynamic dns"""
-
-    if not len(resultList) >= limit:
-        if verbose:
-            print("[+] Dynamic DNS")
-        
-        try:
-            r = requests.get("https://raw.githubusercontent.com/MISP/misp-warninglists/main/lists/dynamic-dns/list.json")
-            dynamicdns = r.json()['list']
-            with open(pathEtc + "/dynamic-dns.json", "w") as write_json:
-                json.dump(r.json(), write_json, indent=4)
-        except:
-            with open(pathEtc + "/dynamic-dns.json", "r") as read_json:
-                dynamicdns = json.load(read_json)["list"]
-
-
-        resultLoc = list()
-
-        domain_replace = [domain[::-1].replace(".", "-", 1)[::-1], domain.replace(".", "-"), domain.replace(".", "_")]
-
-        for ddns in dynamicdns:
-            for replace in domain_replace:
-                d_loc = replace + ddns
-                if d_loc not in resultLoc:
-                    resultLoc.append(d_loc)
-
-        if verbose:
-            print(f"{len(resultLoc)}\n")
-
-        if combo:
-            resultLoc = checkResult(resultLoc, resultList, givevariations, "addDynamicDns")
-            resultLoc = final_treatment(domain, resultLoc, limit, givevariations, keeporiginal, "addDynamicDns")
-            return resultLoc
-
-        resultList = checkResult(resultLoc, resultList, givevariations, "addDynamicDns")
-        resultList = final_treatment(domain, resultList, limit, givevariations, keeporiginal, "addDynamicDns")
-
-    return resultList
-
-
-def numeralSwap(domain, resultList, verbose, limit, givevariations=False,  keeporiginal=False, combo=False):
-    """Change a numbers to words and vice versa"""
-
-    if not len(resultList) >= limit:
-        if verbose:
-            print("[+] Numeral Swap")
-
-        resultLoc = list()
-        loclist = list()
-
-        prefix, domain_without_tld, tld = parse_domain(domain)
-        domainList = [domain_without_tld]
-
-        for name in domainList:
-            for numerals_list in numerals:
-                for nume in numerals_list:
-                    if nume in name:
-                        for nume2 in numerals_list:
-                            if not nume2 == nume:
-                                loc = prefix + name.replace(nume, nume2)
-                                if not loc in resultLoc:
-                                    resultLoc.append(loc)
-            if resultLoc:
-                loclist.append(resultLoc)
-                resultLoc = list()
-
-        if loclist:
-            loclist.append([tld])
-            rLoc = globalAppend(loclist)
-
-            if verbose:
-                print(f"{len(rLoc)}\n")
-
-            if combo:
-                rLoc = checkResult(rLoc, resultList, givevariations, "numeralSwap")
-                rLoc = final_treatment(domain, rLoc, limit, givevariations, keeporiginal, "numeralSwap")
-                return rLoc
-
-            resultList = checkResult(rLoc, resultList, givevariations, "numeralSwap")
-            resultList = final_treatment(domain, resultList, limit, givevariations, keeporiginal, "numeralSwap")
-        elif verbose:
-            print("0\n")
-
-
-    return resultList
-
+### SEPARATOR 1 ###
 
 def catchAll(current_domain):
     import dns.resolver
@@ -1243,7 +190,6 @@ def catchAll(current_domain):
     if is_catch_all:
         return domain_catch_all
     return False
-        
 
 def dnsResolving(resultList, domain, pathOutput, verbose=False, givevariations=False, dns_limited=False, catch_all=False):
     """Do a dns resolving on each variations and then create a json"""
@@ -1333,6 +279,7 @@ def dnsResolving(resultList, domain, pathOutput, verbose=False, givevariations=F
 
     return domain_resolve
 
+### SEPARATOR 2 ###
 
 def formatYara(resultList, domain, givevariations=False):
     """Output in yara format"""
@@ -1447,6 +394,7 @@ def formatOutput(format, resultList, domain, pathOutput, givevariations=False, b
         print(f"Unknown format: {format}. Will use text format instead")
         formatOutput("text", resultList, domain, pathOutput, givevariations)
 
+### SEPARATOR 3 ###
 
 def runAll(domain, limit, formatoutput, pathOutput, verbose=False, givevariations=False, keeporiginal=False, all_homoglyph=False):
     """Run all algo on each domain contain in domainList"""
@@ -1465,7 +413,7 @@ def runAll(domain, limit, formatoutput, pathOutput, verbose=False, givevariation
     return resultList
 
 
-
+# Main file function
 if __name__ == "__main__":
 
     parser = argparse.ArgumentParser()
@@ -1598,8 +546,7 @@ if __name__ == "__main__":
                                 base_result = loc_result
 
                                 if verbose:
-                                    print(f"{len(loc_result)}\n")
-                                    
+                                    print(f"{len(loc_result)}\n")                                 
         elif args.all:
             for algo in algo_list:
                 func = globals()[algo]
